JavaScript security: endspoints and environment variable allow/denylist options (#352)
Adjusted to actual behavior of endpoint filtering, some other adjustments.

tcp:// and ssl:// are still supported but somehow buggy. Updated docs only describe http:// and https:// that can now be used thanks to arangodb/arangodb#11236

h2://, h2s://, srv:// and unix:// are supported as well but not left uncovered in the docs.